+++ subsys/bluetooth/host/Kconfig
@@ -1,6 +1,6 @@
 # Bluetooth LE stack configuration options
 
-# Copyright (c) 2016-2017 Nordic Semiconductor ASA
+# Copyright (c) 2016-2020 Nordic Semiconductor ASA
 # Copyright (c) 2015-2016 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
@@ -540,6 +540,13 @@ config BT_TINYCRYPT_ECC
 	  Connections so that Hosts can make use of those. The option defaults
 	  to enabled for a combined build with Zephyr's own controller, since it
 	  does not have any special ECC support itself (at least not currently).
+
+config BT_HOST_CCM
+	bool "Enable host side AES-CCM module"
+	help
+	  Enables the software based AES-CCM engine in the host. Will use the
+	  controller's AES encryption functions if available, or BT_HOST_CRYPTO
+	  otherwise.
 
 if BT_DEBUG
 config BT_DEBUG_SETTINGS
